Refactor puppeteer tests
