+++ dapps/marketplace/test/utils/_services.js
@@ -10,7 +10,7 @@ import services from '@origin/services'
 
 const headless = process.env.HEADLESS === 'false' ? false : true
 
-export default async function() {
+export default async function srv() {
   return await services({
     ganache: { inMemory: true },
     ipfs: true,
@@ -19,6 +19,7 @@ export default async function() {
     relayer: true,
     graphqlServer: true,
     contractsFile: 'tests',
+    quiet: true,
     extras: async () => {
       const webpackProcess = spawn(
         './node_modules/.bin/webpack-dev-server',
@@ -47,16 +48,16 @@ export default async function() {
       const page = pages[0]
 
       await page.goto('http://localhost:8083')
-      await page.evaluate(
-        () =>
-          new Promise(resolve => {
-            window.transactionPoll = 100
-            // window.localStorage.locale = 'zh_CN'
-            resolve()
-          })
-      )
+      await page.evaluate(() => {
+        window.transactionPoll = 100
+        // window.localStorage.locale = 'zh_CN'
+      })
 
       return { page, browser, webpackProcess }
     }
   })
+}
+
+export async function getPage() {
+  return (await srv()).extrasResult.page
 }
