+++ packages/graphql/src/mutations/token/updateTokenAllowance.js
@@ -1,5 +1,7 @@
 import txHelper, { checkMetaMask } from '../_txHelper'
 import contracts from '../../contracts'
+import createDebug from 'debug'
+const debug = createDebug('origin:updateTokenAllowance:')
 
 async function updateTokenAllowance(_, { token, from, to, value }) {
   let tokenContract = contracts.tokens.find(t => t.id === token)
@@ -18,6 +20,7 @@ async function updateTokenAllowance(_, { token, from, to, value }) {
   value = contracts.web3.utils.toWei(value, 'ether')
   const tx = tokenContract.contractExec.methods.approve(to, value)
   const gas = await tx.estimateGas({ from })
+  debug({ token, from, to, value })
   return txHelper({ tx, from, mutation: 'updateTokenAllowance', gas })
 }
